Run LDraw application after checking/creating of $LDRAWHOME directory
This is a wrapper for all the programs that are part of the
linux-ldraw project.  The system LDraw library is installed in
/usr/share/ldraw To allow users to have their own unofficial
parts, a custom ldconfig.ldr or other custom files, all the
programs are configured to use ~/.ldraw as the LDraw library
path by setting the LDRAWDIR environment variable.

If this directory does not exist, it is created and the
contents of the system-wide ldraw library is symlinked there.

All the linux-ldraw packages are supposed to suffix their
binary name with .bin and create a symlink named with the
original binary name pointing to this script